Moved bin function to OOPAO tools, and removed some comments and prints used when testing the geometric SH.
